phase-3-chatbot/backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend"
version = "1.0.0"
description = "Todo application backend"
authors = [{name = "Developer", email = "dev@example.com"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "better-auth>=0.0.1b11",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
    "email-validator>=2.1.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "config*", "database*", "models*", "routes*", "schemas*", "services*", "middleware*", "auth*"]